backend/requirements.txt
annotated_doc==0.0.4
annotated_types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset_normalizer==3.4.4
click==8.3.1
colorama==0.4.6
opencv-contrib-python==4.13.0.90
exceptiongroup==1.3.1
fastapi==0.128.0
h11==0.16.0
idna==3.11
numpy==2.2.6
pillow==12.1.0
pip==22.3.1
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.21
requests==2.32.5
setuptools==65.5.0
starlette==0.50.0
typing_extensions==4.15.0
typing_inspection==0.4.2
urllib3==2.6.3
uvicorn==0.40.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
numpy>=1.24.0
opencv-contrib-python
python-multipart>=0.0.6
Pillow>=10.0.0
